auth/backend/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.3",
    "dj-database-url>=3.0.0",
    "django-dynamic-filenames>=1.4.0",
    "django-ninja>=1.4.3",
    "django-stubs-ext>=5.2.0",
    "django[argon2]>=5.2",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "ulid-py>=1.1.0",
]

[dependency-groups]
dev = [
    "django-cprofile-middleware>=1.0.5",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "django-stubs[compatible-mypy]>=5.2.0",
    "pydot>=4.0.0",
    "pyparsing>=3.2.3",
]
[tool.mypy]
plugins = ["mypy_django_plugin.main"]
mypy_path = "auth"
exclude = ["migrations"]
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"
